Project Creation and home refactored. Require, underscore and backbone added.

+++ Auth/AuthManager.php
@@ -14,14 +14,20 @@
 	public $email;
 	public $language = "en";
 	
-	private $serverManager = false;
+        private $app;
+        private $db;
+        private $serverManager = false;
 	
 	private $openIdEnabled = true;
 	private $ldapEnabled = true;
 	private $localEnabled = true;
 	
-	function __construct($cfg)
+	function __construct($app, $cfg)
 	{	
+                $this->app = $app;
+                $this->db = $this->app->getDB();
+                $this->cfg = $cfg;
+
 		if(array_key_exists("use_openID", $cfg->settings["security"]))
 		{
 			$this->openIdEnabled = $cfg->settings["security"]["use_openID"] == "true";
@@ -43,7 +49,7 @@
 	  	{
 	  		try
 	  		{
-	   			$this->providers["OPENID"] = new OpenIDProvider("http://test.mlst.net/index.php");
+	   			$this->providers["OPENID"] = new OpenIDProvider($this->app);
 	  		}
 	  		catch(Exception $err)
 	  		{
@@ -79,52 +85,54 @@
         }
   
   	function requestlogin($requestedUrl, $provider = "")
-  	{
-  		global $cfg, $hasManagers, $SITE_ROOT;
-  		
-  		$provider = strtoupper($provider);
-  		
-  		$_SESSION["url"] = "http://{$_SERVER['HTTP_HOST']}{$SITE_ROOT}/" . trim($requestedUrl, '/');
-  		
-  		
-  		if(($provider != "" && $provider != "LOCAL" && array_key_exists($provider, $this->providers)) || count($this->providers) == 1)
-  		{
-  			if($provider == '' && count($this->providers) == 1)
-  			{
-  				$keys = array_keys($this->providers);
-  				$provider = $keys[0];
-  				$_SESSION['provider'] = $provider;
-  			}
-  			return $this->providers[$provider]->requestLogin("http://{$_SERVER['HTTP_HOST']}{$SITE_ROOT}/" . trim($requestedUrl, '/'), !$hasManagers);
-  		}
-  		else
-  		{
-  			global $url, $SITE_ROOT;
-  			$server = trim($_SERVER["HTTP_HOST"], "/");
-  			$root = trim($SITE_ROOT, "/");
-  			$frm =  "<p>Please Choose a Method to login</p>";
-  			if($this->localEnabled)$frm .= "<div class=\"login-provider epicollect\"><h3>EpiCollect Account</h3>" .  $this->providers["LOCAL"]->requestLogin("http://{$_SERVER['HTTP_HOST']}{$SITE_ROOT}/" . trim($requestedUrl, '/'), !$hasManagers) . "</div>";
-  			if($this->openIdEnabled) $frm .= "<div class=\"login-provider google\"><h3>Google/Gmail</h3><a class=\"btn\" href=\"http://$server/$root/$url?provider=OPENID\">Google/Gmail account (OpenID)</a></div>";
-			if($this->ldapEnabled && array_key_exists("ldap_domain", $cfg->settings["security"]) && $cfg->settings["security"]["ldap_domain"] != "")
-			{
-					$frm .= "<a class=\"login-provider\" href=\"http://$server/$root/$url?provider=LDAP\">Windows Account ({$cfg->settings["security"]["ldap_domain"]})</a>";
-			}
-			return $frm;
-  		}
+  	{  		
+            $provider = strtoupper($provider);
+
+            if($requestedUrl === 'login')
+            {
+                $_SESSION["url"] = $this->app->site_root . '/';
+            }
+            else
+            {
+                $_SESSION["url"] = $this->app->site_root . '/' . trim($requestedUrl, '/');
+            }    
+
+
+            if(($provider != "" && $provider != "LOCAL" && array_key_exists($provider, $this->providers)) || count($this->providers) == 1)
+            {
+                if($provider == '' && count($this->providers) == 1)
+                {
+                        $keys = array_keys($this->providers);
+                        $provider = $keys[0];
+                        $_SESSION['provider'] = $provider;
+                }
+                return $this->providers[$provider]->requestLogin($this->app->base_url . trim($requestedUrl, '/'), false);
+            }
+            else
+            {
+                $frm =  "<p>Please Choose a Method to login</p>";
+                if($this->localEnabled)$frm .= "<div class=\"login-provider epicollect\"><h3>EpiCollect Account</h3>" .  $this->providers["LOCAL"]->requestLogin("{$this->app->base_url}/" . trim($requestedUrl, '/'), false) . "</div>";
+                if($this->openIdEnabled) $frm .= "<div class=\"login-provider google\"><h3>Google/Gmail</h3><a class=\"btn\" href=\"{$this->app->base_url}/$requestedUrl?provider=OPENID\">Google/Gmail account (OpenID)</a></div>";
+                if($this->ldapEnabled && array_key_exists("ldap_domain", $cfg->settings["security"]) && $cfg->settings["security"]["ldap_domain"] != "")
+                {
+                    $frm .= "<a class=\"login-provider\" href=\"{$this->app->base_url}/$requestedUrl?provider=LDAP\">Windows Account ({$cfg->settings["security"]["ldap_domain"]})</a>";
+                }
+                return $frm;
+            }
    		
   	}
   	
   	function setEnabled($uid, $enabled)
   	{
-  		global $db;
-  		
-  		$enabled = $enabled ? "1" : "0";
-  		if($uid != $this->getEcUserId())
-  		{
-  			$qry = "UPDATE user SET active = $enabled where idUsers = $uid";
-  			return $db->do_query($qry);
-  		}
-  		return false;
+            global $db;
+
+            $enabled = $enabled ? "1" : "0";
+            if($uid != $this->getEcUserId())
+            {
+                    $qry = "UPDATE user SET active = $enabled where idUsers = $uid";
+                    return $db->do_query($qry);
+            }
+            return false;
   	}
   	
   	function resetPassword($uid)
@@ -155,15 +163,14 @@
   
   	function callback($provider = "")
   	{
-  		global  $cfg, $db, $SITE_ROOT, $url;
                 
   		if( $this->isLoggedIn() ) 
   		{
-  			header("location: http://{$_SERVER["HTTP_HOST"]}{$SITE_ROOT}/{$_SESSION["url"]}"); return;
+  			header("location: {$this->app->base_url}/{$_SESSION["url"]}"); return;
   		}
   		
   		if(!array_key_exists($provider, $this->providers)) {
-  			header("location: http://{$_SERVER["HTTP_HOST"]}{$SITE_ROOT}/");
+  			header("location: {$this->app->base_url}/");
 
   		}
   		  		
@@ -185,9 +192,9 @@
 	  			$this->language = $this->providers[$provider]->language;
   			}
   			
-  			$res = $db->do_query($sql);
-  			if($res !== true) die($res . "\n" . $sql);
-  			while($arr = $db->get_row_array())
+  			$res_get_user = $this->db->do_query($sql);
+  			if($res_get_user !== true) die($res . "\n" . $sql);
+  			while($arr = $this->db->get_row_array())
   			{
   		
   				if($arr["active"])
@@ -197,24 +204,24 @@
   				else 
   				{
   					flash ("Account is disabled", "err");  	
-  					header("location: {$_SESSION["url"]}");
+  					header("location: {$this->app->base_url}/{$_SESSION["url"]}");
   					return;
   				}
   			}
   			if($provider != "LOCAL" && !$uid)
   			{
   				$sql = "INSERT INTO user (FirstName, LastName, Email, details, language, serverManager) VALUES ('{$this->firstName}','{$this->lastName}','{$this->email}','" . $this->providers[$provider]->getCredentialString() . "','{$this->language}', " . (count($this->getServerManagers()) == 0 ?  "1" : "0") . ")";
-  				$res = $db->do_query($sql);
+  				$res = $this->db->do_query($sql);
   				if($res !== true) die($res);
-  				$uid = $db->last_id();
+  				$uid = $this->db->last_id();
   				if(!$uid) die("user creation failed $sql");
   			}
   			
   			$dat = new DateTime();
-  			$dat = $dat->add(new DateInterval("PT{$cfg->settings["security"]["session_length"]}S"));
+  			$dat = $dat->add(new DateInterval("PT{$this->cfg->settings["security"]["session_length"]}S"));
   			$sql = "INSERT INTO ecsession (id, user, expires) VALUES ('" . session_id() . "', $uid, " . $dat->getTimestamp() . ");";
   			
-  			$res = $db->do_query($sql);
+  			$res = $this->db->do_query($sql);
   			if($res !== true && !preg_match("/Duplicate Key/i", $res)) die($res . "\n" . $sql);
   
   			header("location: {$_SESSION["url"]}");
@@ -232,18 +239,16 @@
   				if($_SESSION["tries"] < 6) $_SESSION["tries"]++;
   			}
   			//sleep($_SESSION["tries"] * $_SESSION["tries"]);
-  			global $SITE_ROOT;
-  			header("location: http://{$_SERVER["HTTP_HOST"]}{$SITE_ROOT}/login.php");
+  			
+  			header("location: {$this->app->base_url}/login.php");
   		}
   	}
   	
   	function logout($provider = "")
   	{
   		//if(!array_key_exists($provider, $this->providers)) return false;
-  		global $db;
-  		if(!$db) $db = new EpiCollectDatabaseConnection();
-  		  		
-  		$res = $db->do_query("DELETE FROM ecsession WHERE id = '" . session_id() . "'");
+  		  		  		
+  		$res = $this->db->do_query("DELETE FROM ecsession WHERE id = '" . session_id() . "'");
   		if(!$res) die("$res - $sql");
   		$_SESSION['provider'] = null;
   		$params = session_get_cookie_params();
@@ -257,40 +262,28 @@
   
   	function isLoggedIn()
   	{
-  		global $db;
-  		if(!$db)
-  		{
-  			try 
-  			{
-  				$db = new EpiCollectDatabaseConnection();
-  			}
-  			catch(Exception $e)
-  			{
-  				return false;
-  			}	
-  		}
-  		
-  		if(!$db->connected) return false;
+  		if(!$this->db || !$this->db->connected) return false;
+                
   		$dat = new DateTime();
   		$qry = "DELETE FROM ecsession WHERE expires < ". $dat->getTimestamp();
   		 		
-  		$res = $db->do_query($qry);
+  		$res = $this->db->do_query($qry);
   		if($res !== true) return false;
   		
   		$this->user = false;
   		
   		$qry = "select user, firstName, lastName, email, serverManager from ecsession left join user on ecsession.user = user.idUsers WHERE ecsession.id = '" . session_id() ."'"; 
-  		$res = $db->do_query($qry);
+  		$res = $this->db->do_query($qry);
   		if($res !== true) die($res . "\n" . $qry);
   		
-  		while ($arr = $db->get_row_array()){ 
+  		while ($arr = $this->db->get_row_array()){ 
   			$this->user = $arr["user"]; 
   			$this->firstName = $arr["firstName"];
   			$this->lastName = $arr["lastName"];
   			$this->email = $arr["email"];
   			$this->serverManager = $arr["serverManager"];
   		}
-		$db->free_result();
+		$this->db->free_result();
    		return $this->user !== false;
   	}
   
@@ -301,11 +294,10 @@
   	
   	function makeServerManager($email)
   	{
-  		global $db;
   		$qry = "select serverManager from user WHERE email = '$email'";
   		$res = $db->do_query($qry);
   		$r=0;$u=0;
-  		while ($arr = $db->get_row_array())
+  		while ($arr = $this->db->get_row_array())
   		{
   			$u++;
   			$r += $arr["serverManager"];
@@ -321,28 +313,26 @@
   	
   	function removeServerManager($email)
   	{
-  		global $db;
   		$qry = "UPDATE user SET serverManager = 0 WHERE email = '$email'";
-  		if($db->do_query($qry) !== true) die("oops");
+  		if($this->db->do_query($qry) !== true) die("oops");
   	}
   	
   	function getServerManagers()
   	{
-  		global $db;
   		try{
   		
   			$men = array();
-  			if($db)
+  			if($this->db)
   			{
 		  		$qry = "SELECT firstName, lastName, Email FROM user WHERE serverManager = 1 and active = 1";
-		  		$res = $db->do_query($qry);
+		  		$res = $this->db->do_query($qry);
 		  		if($res !== true) throw new Exception(sprintf('MySQL error :  %s last successful query was %s',$res, $db->lastQuery));
 				
-		  		while($arr = $db->get_row_array())
+		  		while($arr = $this->db->get_row_array())
 		  		{
 		  			array_push($men, $arr);
 		  		}
-		  		$db->free_result();
+		  		$this->db->free_result();
   			}
   			//
 	  		return $men;
@@ -355,12 +345,11 @@
   	
   	private function populateSesssionInfo()
   	{
-	   $db = new EpiCollectDatabaseConnection();
 	   $qry = "SELECT idUsers as userId, FirstName, LastName, Email, language FROM user WHERE openId = '{$_SESSION['openid']}'";
-	   $err = $db->do_query($qry);
+	   $err = $this->db->do_query($qry);
 	   if($err === true)
 	   {
-			if($arr = $db->get_row_array())
+			if($arr = $this->db->get_row_array())
 			{
 	 			foreach(array_keys($arr) as $key)
 	 			{
@@ -395,9 +384,8 @@
 	  
 	  function getUsers($order = "FirstName", $dir = "asc")
 	  {
-                    global $db, $log;
                     $query = "SELECT idUsers as userId, FirstName, LastName, Email, active FROM user ORDER BY $order $dir";
-                    $res = $db->do_query($query);
+                    $res = $this->db->do_query($query);
                     if(!$res === true)
                     {
                             $log->write("err", $res);
@@ -406,7 +394,7 @@
                     else
                     {
                             $ret = array();
-                            while($arr = $db->get_row_array())
+                            while($arr = $this->    db->get_row_array())
                             {
                                     array_push($ret, $arr);
                             }
